+++ apis/eks/v1alpha1/nodegroup_types.go
@@ -71,7 +71,7 @@ type NodeGroupParameters struct {
 	// disk size is 20 GiB.
 	// +immutable
 	// +optional
-	DiskSize *int64 `json:"diskSize,omitempty"`
+	DiskSize *int32 `json:"diskSize,omitempty"`
 
 	// The instance type to use for your node group. Currently, you can specify
 	// a single instance type for a node group. The default value for this parameter
@@ -194,23 +194,23 @@ type NodeGroupScalingConfig struct {
 	// is expected to manage the desired size of the node group. If not set, the initial
 	// desired size will be the configured minimum size of the node group.
 	// +optional
-	DesiredSize *int64 `json:"desiredSize,omitempty"`
+	DesiredSize *int32 `json:"desiredSize,omitempty"`
 
 	// The maximum number of worker nodes that the managed node group can scale
 	// out to. Managed node groups can support up to 100 nodes by default.
 	// +optional
-	MaxSize *int64 `json:"maxSize,omitempty"`
+	MaxSize *int32 `json:"maxSize,omitempty"`
 
 	// The minimum number of worker nodes that the managed node group can scale
 	// in to. This number must be greater than zero.
 	// +optional
-	MinSize *int64 `json:"minSize,omitempty"`
+	MinSize *int32 `json:"minSize,omitempty"`
 }
 
 // NodeGroupScalingConfigStatus is the observed scaling configuration for a node group.
 type NodeGroupScalingConfigStatus struct {
 	// The current number of worker nodes for the managed node group.
-	DesiredSize *int64 `json:"desiredSize,omitempty"`
+	DesiredSize *int32 `json:"desiredSize,omitempty"`
 }
 
 // NodeGroupObservation is the observed state of a NodeGroup.
